Move Channel low-level bits to own module.
