+++ http2-client.cabal
@@ -19,6 +19,7 @@ library
                      , Network.HTTP2.Client.FrameConnection
                      , Network.HTTP2.Client.Helpers
                      , Network.HTTP2.Client.RawConnection
+  other-modules:       Network.HTTP2.Client.Channels
   build-depends:       base >= 4.7 && < 5
                      , async
                      , bytestring
